backend/requirements.txt
# Django and REST Framework
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.0

# Database
psycopg2-binary>=2.9.0
django-environ>=0.11.0

# Authentication and Security
djangorestframework-simplejwt>=5.3.0
django-ratelimit>=4.1.0
cryptography>=41.0.0

# Blockchain and Web3
web3<7,>=6.0.0
eth-account==0.11.0
eth-utils==3.0.0
eth-abi==5.0.0

# Encryption
gmpy2==2.2.1
pycryptodome==3.23.0

# Face Recognition (Self-hosted)
face-recognition>=1.3.0
dlib>=19.22.0 # Required by face-recognition

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.1.0

# API and HTTP
requests>=2.31.0
celery>=5.3.0
redis>=5.0.0

# Development and Testing
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Utilities
python-decouple>=3.8
django-extensions>=3.2.0 